- mean leaf depth computation added - experiments for conjecture about relation of backpropagating means and values MCTS added - SM-OOS variants dropping the tree and using in-match probabilities implemented

+++ src/cz/agents/gtlibrary/algorithms/mcts/selectstrat/BasicStats.java
@@ -12,7 +12,7 @@ import java.io.Serializable;
  * @author Vilo
  */
 public class BasicStats implements Serializable {
-	private int n = 0;
+	private long n = 0;
 	private double oldM;
 	private double newM;
 
@@ -35,7 +35,7 @@ public class BasicStats implements Serializable {
 		return (n > 0) ? newM : 0.0;
 	}
 
-	public int getNbSamples() {
+	public long getNbSamples() {
 		return n;
 	}
